# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Supabase Modular Auth API
  version: 1.0.0
  description: |
    Contract for the stateless Express API used by the generic SaaS starter.
    Authentication credentials are managed exclusively by the backend in
    HttpOnly cookies. Browser clients must include credentials and must send the
    double-submit CSRF cookie value in `X-CSRF-Token` for every unsafe request.
  license:
    name: ISC
    identifier: ISC
servers:
  - url: http://localhost:3000
    description: Local Express API
...
paths:
...
  /billing/webhook:
    post:
      operationId: verifyStripeWebhook
      summary: Verify a Stripe Checkout webhook
      description: |
        Verifies the Stripe-Signature header against the exact raw request body.
        The smoke test logs verified events but does not fulfill a product purchase.
      tags: [Billing]
      security:
        - stripeSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/StripeWebhookVerified"
        "400":
          $ref: "#/components/responses/BadRequest"
        "408":
          $ref: "#/components/responses/RequestTimeout"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: auth_token
      description: Default HttpOnly access-token cookie; the configured name may differ.
    refreshCookie:
      type: apiKey
      in: cookie
      name: auth_token_refresh
      description: Default HttpOnly rotating refresh-token cookie; the configured name may differ.
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe webhook signature verified against the endpoint secret and raw body.
...
  headers:
    RequestId:
      description: Correlation identifier assigned to the request.
      schema:
        type: string
    SetCookie:
      description: One or more cookies issued or rotated by the backend.
      schema:
        type: string
  responses:
...
    StripeWebhookVerified:
      description: Stripe signature verified and event inspected.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StripeWebhookReceiptResponse"
...
    BadRequest:
      description: Request validation failed.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RequestTimeout:
      description: Request processing timed out.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: A global, auth, sensitive-auth, or admin rate limit was exceeded.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        RateLimit:
          description: Draft-7 RateLimit fields emitted by express-rate-limit.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: An unexpected server error occurred.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: A required upstream service is temporarily unavailable; auth cookies are preserved.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Set-Cookie:
          $ref: "#/components/headers/SetCookie"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        data: {}
...
    ErrorResponse:
      type: object
      required: [success, error, message]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
          enum:
            - AUTH_FAILED
            - INVALID_CREDENTIALS
            - UNAUTHORIZED
            - EMAIL_NOT_VERIFIED
            - FORBIDDEN
            - USER_EXISTS
            - USER_NOT_FOUND
            - VALIDATION_ERROR
            - INVALID_INPUT
            - RATE_LIMITED
            - RATE_LIMIT_EXCEEDED
            - CONNECTION_FAILED
            - SERVICE_UNAVAILABLE
            - REGISTRATION_FAILED
            - INVALID_TOKEN
            - TOKEN_EXPIRED
            - INTERNAL_ERROR
            - CSRF_TOKEN_MISSING
            - CSRF_TOKEN_INVALID
            - REQUEST_TIMEOUT
        message:
          type: string
        details:
          type: array
          description: Present only in development for validation failures.
          items:
            $ref: "#/components/schemas/ValidationErrorDetail"
        retryAfter:
          type: number
          minimum: 0
          description: Retry delay reported by some rate-limit responses.
    ValidationErrorDetail:
      type: object
      required: [code, message, path]
      properties:
        code:
          type: string
        message:
          type: string
        path:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
...
    StripeWebhookReceiptResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [checkoutCompleted, eventId, eventType, paymentStatus]
              properties:
                checkoutCompleted:
                  type: boolean
                eventId:
                  type: string
                eventType:
                  type: string
                paymentStatus:
                  type: [string, "null"]
...